refactor: refactor application structure
