+++ .eslintrc.json
@@ -1,0 +1,17 @@
+{
+    "env": {
+        "browser": true,
+        "es2021": true
+    },
+    "extends": "eslint:recommended",
+    "parserOptions": {
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "rules": {
+        "no-undef": "off",
+        "semi": "error",
+        "space-before-function-paren": ["error", "never"],
+        "space-before-blocks": ["error", { "functions": "always", "keywords": "always", "classes": "always" }]
+    }
+}
