+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "start": "node app.mjs"
+    "start": "node src/app.mjs"
   },
   "dependencies": {
     "@grpc/grpc-js": "^1.8.17",
